Fixes compile warnings
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@32938 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/os2/fontutil.cpp
@@ -413,7 +413,7 @@ void wxFillLogFont(
         }
         if (*pflId == 0L)
             *pflId = 1L;
-        for(unsigned long LCNum = 0; LCNum < lNumLids; LCNum++)
+        for(unsigned long LCNum = 0; LCNum < (unsigned long)lNumLids; LCNum++)
             if(alIds[LCNum] == *pflId)
                ++*pflId;
         if(*pflId > 254)  // wow, no id available!
